updated yaml plugins to lazily import base plugin classes as well
